playbooks: Fix `services` playbook
Because not all node facts were gathered when only running the
`services` playbook, the `kube_prometheus` role failed since it couldn't
find the IP addresses of all `etcd` and `k8s-cluster` nodes.

This is somewhat of a band-aid which ensures facts *are* gathered. Maybe
adding this to the code which 'requires' these values (i.e. somewhere in
the `kube_prometheus` role) would be better...

Fixes: #120
See: https://github.com/scality/metal-k8s/issues/120