docs: Move introduction into a separate document
Because of the change in e0b28606f7253, the introduction text would no
longer be rendered in the PDF documentation, which isn't the intent.
Moving it to a separate document instead.

However, by doing so, the index page of the HTML build would only
contain the TOC tree. As such, replicating part of it to that page.

See: e0b28606f7253a23f2c1d66e47dddb88d8cf70f1