salt: don't check for utilities in `__virtual__`
If we check for utility programs and their functionality in
`__virtual__` of a module, the module will be unavailable when it's
loaded.

However, many of our states install or configure those utilities, and
then want to use a Salt module or state depending on them, which would
then fail.

This commit removes those checks from the various modules. Instead an
invocation of the program would fail at execution time.