Split and factorize Makefile for packages
In order to facilitate the generation of new packages for supporting
Python 3, we factorize the rules building process using templates with a
combination of `$(call ...)` and `$(eval ...)`.

This permits to achieve the desired clarity in `packages/packages.mk`
and `packages/repositories.mk`, at the cost of much lower readability in
the rule templates in `packages/Makefile`.

Hopefully, this painful example will serve as an argument in favor of
replacing GNU Make with a more powerful alternative.