salt: introduce `BootstrapConfiguration`
This commit introduces a `BootstrapConfiguration` file which is supposed
to be present on a bootstrap system. This file will be read by an
'external pillar' to populate pillars.

This removes the arbitrary hard-coded IP addresses from the pillar files
coming with the distribution.

Fixes: #776
See: https://github.com/scality/metalk8s/issues/776