build: introduce doit
Our new build system will be based on `doit`[1] instead of GNU Make.
This will certainly be more maintainable.

This commit adds the necessary requirements.txt to install doit (as well
as mypy (type annotations FTW) and pylint as dev dependencies).

It also update the gitignore to ignore the cruft that we don't want
to track in Git.

[1]: http://pydoit.org/

Refs: #708
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>