build/vagrant: add a task to run `vagrant`
`make vagrantup` is now superseded by `doit vagrant`.

We explicitely set `uptodate`[1] to False as this task can never be
up-to-date (i.e. each time you run it you want it to spawn an
environment).

Note that this task should depend on tasks that populate the ISO, but we
can't do it yet (because those tasks aren't available).

This will be done in later commits, as we keep migrating the Makefile.

[1]: http://pydoit.org/dependencies.html#uptodate

Refs: #708
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>