build/target: add a thin object layer on tasks
Using task as dict is convenient and easy, but for complex task it's
likely that you'll want to combine some fields or do some computation
(e.g. infer the path where to store a package from its name and version,
…) and to do so you'll need an object.

There is also some default behavior that we'll want to apply (e.g.
allowing task to "revert" their actions, which is done by setting
`clean` to True in most of the cases) and a class is also convenient
here.

That's enough reason to create this `Target` abstract classes.

Another thing to consider is that while simple target will probably be
produced by a single task, more complex pone (such as creating a
package) will be more involved and will likely require several tasks to
be accomplished: enter `AtomicTarget` and `CompositeTarget`.

Refs: #708
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>