build/utils: add an helper for a nicer output
`title` are used to represent a task in doit's output.
Frequently we will want to display the command (or a symbolic command)
followed by its first target (especially since we will usually only have
one target in fact), e.g.:
- "MKDIR    _build/foo"
- "MKISOFS  bar.iso"
- "RENDER   baz.sls.in"
- …

Since it's a common need, it deserves an helper function.

Note that in order to have a nicer output, we need two things:
- alignment: to do so, we define a max length for the symbolic command
- avoid overlong line: absolute paths are converted to path relative to
  the build root

Refs: #708
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>