build: add a task to create the build root
As it only need to be run once, we use the helper `run_once`[1].

Note the double `_`: it makes the task private.

By default, the name of the task is everything after `task_` AND tasks
whose names starts with an underscore are private/hidden so
`task__build_root` creates the task `_build_root` which is private.

Using private tasks avoid cluttering the output of `doit list` with
rarely used tasks (and if you really want to see them, you can still use
`doit list -p` or `doit list --private`).

[1]: http://pydoit.org/uptodate.html#run-once

Refs: #708
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>