build/iso: add a module to build the final ISO
For now we only create the root directory.

This module will grow as we continue to convert the Makefile targets to
doit tasks.

We configure `doit` to make `iso` the default task (like `iso` is the
default target in our Makefile).
And since we're addind configuration, we enable by default the cleanup
of the task dependencies: by doing that, when we simply type
`doit clean iso` to the task and its dependencies.

Without this option, we would have to type `doit clean -c iso` (
or `doit clean --clean-dep iso`), otherwise nothing would be cleaned
(because `iso` is a task-group, it produces nothing by itself and thus
has nothing to clean).

Refs: #708
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>