build/constants: load version information
These information will be required to generate `product.txt` and
`bootstrap.sh`.

We inject them as globals, so that they can be used like the other
constants.

We can't include them directly in `config.py` because I think that the
CI relies on the VERSION file, and since we don't want to duplicate this
information then we have to import it.

Refs: #708
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>