build/target: add the FileTarget base class
Having a task producing a single output file is common enough (rendering
template, computing a checksum, building a container image) to deserver
a base class.

This base class does the bare minimum: setting the target to a single
element list and exposing a property (`destination`)  that avoids/hides
the ugly `targets[0]`.

Refs: #708
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>