build/target: add a generic task for template rendering
Another common needs during a build is to generate files from template
that contains placeholder that can only be filled at build time (e.g.
bootstrap.sh, top.sls, …).

To do so, we add a new generic task which is based on the standard
Python template engine (less powerful than Jinja, for sure but for now
our needs are pretty limited so there is no reason to add an
extra-dependency for that).

Note that by default, the placeholder delimiter is `$` which is a pretty
bad choice when you templatize shell script (you'll have to escape a lot
of legitimate `$`).
We will use `@@` instead (so instead of writing `@VERSION@` we will
write `@@VERSION`.

Refs: #708
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>