build/iso: add a task to populate the ISO tree
Populating the ISO tree with all the required files (container images,
software repositories, Salt states, …) will be the bulk of the work done
by our build system, so we can naturally expect this to span over
several tasks.

Now there are some cases where one could want to prepare everything for
the ISO without actually building an ISO: e.g. the task `vagrant` don't
need to the final ISO, just deploying the files is enough (they are then
"mounted" into the environment).

This commit introduce a task-group[1] that will gathers every task that
create files into the ISO file tree.

`make all-local` is superseded by the task-group `_iso_populate` (unlike
`all-local`, this one is a private task because a human is unlikely to
run this task directly: either we will build the ISO or we will run
`vagrant`, those tasks will run `_iso_populate` as needed (and if one
really wants to run it, then `doit _populate_iso` is still an option)).

[1]: http://pydoit.org/dependencies.html?#groups

Refs: #708
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>