build/iso: add a task to generate the bootstrap.sh
Using a brand new helper `TemplateFile`, it is now easy to render the
bootstrap script `bootstrap.sh` from its template `bootstrap.sh.in`.

Note that this new task depends on the VERSION file, because we use the
SHORT_VERSION (which is computed from this file) as value during the
rendering, consequently if VERSION changes, we must re-render the
script.

Refs: #708
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>