build/iso: add a task to build the final ISO
Note that the dependencies computation for this task must be delayed
until the ISO root is populated (hence the use of `create_after`),
because we wants to depends on all the files in the ISO tree and this
can only be computed after populating the ISO tree.

Thanks to delayed task[1] AND ou task-group `_populate_iso` (which is a
savior here, as we can only specify one single task for `executed`)
introduced earlier, it's relatively easy to do.

[1]: http://pydoit.org/task_creation.html#delayed-task-creation

Refs: #708
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>