build/coreutils: implement sha256sum
This commit adds a pure-Python implementation of the `sha256sum` command
(and an associated task template to ease its reuse).

The advantages of using a Python function instead of calling the
original binary are:
- portability (even though `sha256sum` is pretty ubiquitous).
- avoid a fork.
- no need to spawn a shell (`shell=True` for `subprocess.Popen`) to use
  redirections (e.g. `sha256sum foo > bar`), which is a good things to
  avoid.

Note that we take care to use incremental hashing, because loading the
whole file into memory may not be a good idea (we will checksum an ISO
and an ISO can take several Go of memory…).

Also, we can expect that the Python dev are not totally dumb that
`sha256` is in fact calling C code so the perf shouldn't be abysmal.

Anyway, let's verify it with a `hash.py` containing:

    coreutils.sha256sum([Path("_build_ref/metalk8s.iso")], Path("/tmp/sha256sum"))

We get:

    % hyperfine 'sha256sum _build_ref/metalk8s.iso > /tmp/sha256sum' './hash.py'
    Benchmark #1: sha256sum _build_ref/metalk8s.iso > /tmp/sha256sum
    Time (mean ± σ):      2.110 s ±  0.003 s    [User: 1.987 s, System: 0.121 s]
    Range (min … max):    2.106 s …  2.115 s    10 runs

    Benchmark #2: ./hash.py
    Time (mean ± σ):      2.167 s ±  0.033 s    [User: 2.016 s, System: 0.146 s]
    Range (min … max):    2.148 s …  2.258 s    10 runs

    Summary
    'sha256sum _build_ref/metalk8s.iso > /tmp/sha256sum' ran
        1.03 ± 0.02 times faster than './hash.py'

Looks pretty acceptable to me :)

And in fact, timing are even closer if you doesn't account for the
startup time of the Python interpreter.

    % python3 -m timeit -r 10 -u msec \
        -s 'from pathlib import Path' \
        -s 'from buildchain import coreutils' \
        'coreutils.sha256sum([Path("_build_ref/metalk8s.iso")], Path("/tmp/sha256sum"))'
    1 loop, best of 10: 2.11e+03 msec per loop

(2110 msec => 2.11s => same time as the `sha256sum` binary)

So yeah, "Looks pretty acceptable to me."² :p

Refs: #708
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>