build/coreutils: implement gzip
For the same reasons we've reimplemented `sha256sum`, we provide a
Python equivalent to `gzip`.

We use 6 instead of 9 as default compression level, because that's what
the `gzip` command does (cf. man page).

Unlike `sha256sum`, we don't create a generic task for this one because
it will be used as an action inside existing task instead of being a
standalone task (note that if we ever change our mind, this will be easy
enough to add).

Refs: #708
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>