build/target: add a generic task to pull images
With this task, we can pull container images from online registry and
save them (optionally compressed) locally to embed them in our ISO.

A notable change over the Makefile/shell approach is that here we're
first saving the uncompressed image on disk AND then compressing it
wherease the Makefile use on-the-fly compression (thanks to the pipe).

Is our approach harmful for the performance? Let's see, with a `zip.py`
containining:

    subprocess.run(['docker', 'save', 'salt-master:2018.3.3-1', '-o', 'img.tar'])
    coreutils.gzip(Path('img.tar'))

We get:

    % hyperfine 'docker save salt-master:2018.3.3-1 | gzip > img1.gz' './zip.py'
    Benchmark #1: docker save salt-master:2018.3.3-1 | gzip > img1.gz
    Time (mean ± σ):     18.349 s ±  0.436 s    [User: 14.851 s, System: 0.514 s]
    Range (min … max):   17.984 s … 19.179 s    10 runs

    Benchmark #2: ./zip.py
    Time (mean ± σ):     18.140 s ±  0.478 s    [User: 13.995 s, System: 0.515 s]
    Range (min … max):   17.407 s … 19.030 s    10 runs

    Summary
    './zip.py' ran
        1.01 ± 0.04 times faster than 'docker save salt-master:2018.3.3-1 | gzip > img1.gz'

Seems like they're pretty equivalent (with a lil' speed advantage for us
this time, maybe because we fork once where the shell solution fork
twice).

Refs: #708
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>