build/image: add a module to put images on the ISO
For now we only create the root directory for the container images on
the ISO.

This module will grow as we continue to convert the Makefile targets to
doit tasks.

Because we will have several tasks (mkdir, pull, build) in this module
we create a top-level task-group to gather them. And we make this group
a public task as ones may want to build/download the images only.

Also, considering that this task will put files into the ISO tree, it
must be added as dependency to `_iso_populate`.

Refs: #708
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>