build/target: add a generic task to build images
With this task, we can build container images from a local Dockerfile
and (optionally) save them (compressed) on the disk.

If the image is not saved on disk, a dummy empty file is created to
in the build tree to keep track of it.

Note that for now, if you copy files into your images, you have to add
them manually as `file_dep`. In the future we may "parse" the
Dockerfile, looking for `ADD` and `COPY` statements in order to
automagically populate `file_dep`.

Refs: #708
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>