build/target: workaround a cleaning "bug"
Weirdly enough it seems that when `doit` clean targets it does so from
the beginning to the end whereas I expected it to delete in reverse
order.

Consequently my nice ordering of the directories target (from the root
to the leaves) blow up when cleaning (because it starts `rmdir` from the
root, which is obviously not empty).

That being said, it's easy to workaround: just reverse the sorting order
and the iteration order at creation time.

I'll probably open an issue upstream to clarify if it's a bug or not.

Refs: #708
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>