build/salt: add a task to deploy the Salt tree
Using a generic `FileTree` target, we create two tasks to deploy the
Salt tree (one for the `/pillar` and the other for `/salt`).

Note that we take advantage of the `FileTarget` support to render
`top.sls` and downloading the pause container image on the fly, during
the tree copy.

Like the `image` task, this one expose a public task-group and is
referenced as a depedency of `_iso_populate` as it deploys files under
the ISO root

Refs: #708
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>