build/package: add a module to put repo on the ISO
This is the last big chunk of work to complete the build: we need to
create the software repository on the ISO.

This last step involves several steps and this commit is the first of a
series to implement the packaging.

For now, we:
- build the container image that will be used for the packaging
  operations (downloading packages, building repositories, …).
- create the root directories for the software packages and repositories

Tough work will be done in following commits.

Refs: #708
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>