build/target: add a generic task to build packages
This target abstract all the gory details involved in package creation:
- create the directory for the package
- generate a .meta from the .spec
- download the source files using the URLs in the .meta
- build a SRPM containing the source files

We only build a SRPM here, the RPM depends on the repository and thus
will be built when we build the repository.

Because we manually declare the source file (instead of relying on the
content of the .meta file) we check that every declared source file has
an entry in the .meta file and bail out if it's not the case.
(BTW, note that `curl` is advantagely replaced by `urllib.request` :-))

Also, since all of the command using the build container share a common
set of options (like mounting the entrypoint script :p), we define them
in `constants.py` to avoid duplication.

Refs: #708
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>