build/target: add a generic task to build repo
This target abstract all the gory details involved in repository
creation.

We have two cases:
- repository that contains already existing packages
- repository that contains locally built packages

In the first case, the bulk of the work was already done by the
`_download_packages` task (it has already built the repositories around
the packages we downloaded): we only have to generate the `repomd.xml`.

In the second case, we first have to create the directories for the
repository, then build a RPM from a SRPM for each packages and finally
we can generate the `repomd.xml`.

Refs: #708
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>