build: add a wrapper script around `doit`
In order to avoid bothering the user with spawning a virtual environment
(and having to maintain it up-to-date with our requirements.txt, which
is error prone!) we introduce a minimalist shell script wrapper that
installs `doit` in a virtualenv, taking care to update it if necessary,
and then invokes `doit` with all the command line arguments passed
through (it's like using `doit` directly, which is quite convenient).

We also update the "documentation" accordingly.

Refs: #708
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>