build/targets: fix images' tags for containerd
Images loaded into containerd have more constraints, such as no
compression and also a different format for the tag name.

Instead of adding yet another parameters to control that, let's have a
single boolean to indicate if the image should be loadable in containerd
(and if yes, we do what's necessary).

Refs: #708
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>