Pass ISO root as pytest command-line argument
The previous approach assumed a local build environment to read version
information from the generated "product.txt" file, from which it
inferred the ISO mountpoint.
We found that it was easier to simply pass the mountpoint as a CLI
argument to `pytest`, since it should be a configurable value anyway.

Issue: GH-781