Define kubeconfig fixture
This allows one to run tests with the Python Kubernetes client while
not on the bootstrap node.
The current approach uses the generated `admin.conf`, and modifies a few
values with command-line arguments (these modifications are required by
our CI execution context, where the test runner does not belong to the
control-plane network, which in itself should be fixed).

Issue: GH-781