Use label instead of name for exec in pod
We have a BDD scenario attempting to `kubectl exec` in a static Pod (in
our case, the `salt-master`). However, this implies we know the hostname
of where the Pod runs (since its name depends on it). To work around
this problem, we retrieve the Pod's name dynamically using a label
filter.

Issue: GH-781