build: fix dependencies for MacOS
`pip-compile` freeze the dependencies, but it doesn't handle the
platform-specific dependencies (it will uses the ones used by the
platform used to run `pip-compile`).

Consequentely, a `requirements.txt` generated on Linux may not work on
MacOS.

To avoid this issue (and to avoid having the user to run `pip-compile`
themselves) we now generate one `requirements.txt` per platform and let
`doit.sh` pick the right one (we already need to detect MacOS for the
`stat` command anyway).

Closes: #841
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>