Use macros to build image names
We move the image name macros to metalk8s.registry, instead of with
the control-plane.
We also use the Kubernetes version info from configuration in defaults.

Currently, we import such macros both in the rendering SLS file and in
the templates that will use the generated image names.

Issue: GH-856