build: fix lint-python in tox
When linting Python code we also need to select the proper
requirements.txt, to do so we rely on an environment variable (OSTYPE).

Since we now depends on this environment variable, we need to pass it
from doit when calling `tox -e lint-python` and this can be done by
using a `CmdAction` (which only works with string, hence the
join/map/quote dance).

Refs: #842
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>