Use PTY to get build step output in CI
It seems the execution of `./doit.sh` in CI fails to redirect its output
for the Buildbot worker to catch it. This may be due to buffering
settings, since the `doit` output is much more terse than what `make`
used to print.

In order to prevent this task from being terminated due to inactivity,
we use a pseudo-terminal to run this step, which apparently fixes the
issue.

Issue: GH-867