salt: store CA minion in `BootstrapConfiguration`
Add a new field in the `BootstrapConfiguration` to specify the Salt
minion to use as a CA server, instead of mine-based guesswork.

Also remove the checks for 'existing CAs', which don't really make
sense: in a way, multiple servers may be in a CA 'state', as long as
only one of them (in our case, the one from the configuration) is being
used.

Over time, we could implement an orchestration playbook to 'migrate' a
CA from one node to another.

Fixes: #632
See: https://github.com/scality/metalk8s/issues/632