doc: introduce a good practice guide for Python
This commit is a first draft to document a set of common good practice
to follow when writing Python code for MetalK8s.

Having consistent and homogeneous code is important and can be achieved
by adhering to a common coding style/good practice.

While this can, and must, be enforced automatically by tools such as
linter and code formatter, some recommandations are too high level to be
automatically verified and thus will be checked during review.

"Rules" that can't be enforced automatically must be documented, that
way:
- new developpers can know about it
- comments in review can refer to it (instead of debating or repeating
the same arguments over and over)

And if some pattern is repeatedly pointed in review, then the document
can be updated to add a new "rule".

Closes: #948
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>