salt: change salt-master config generation
In order to support several `saltenv` at the same time (e.g. during an
upgrade), we now generate the configuration file with `file.serialize`
and `merge_if_exists` set to `True`.
With that change, when you re-run the state with a new `saltenv`,
instead of overwriting the existing one it will add a new entry in the
dict.

Refs: #800
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>