bootstrap: retrieve correct control-plane IP
When the API server VIP changes land, a node may have multiple addresses
within the control-plane network. However, the registry will still be
bound to the non-VIP IP. As such, we need to make sure we use the
'correct' control-plane IP of the node, not any random one within that
network.