salt: don't calculate Pillar top based on cluster
In 13485d2141, changes were made to the way Pillar data is calculated,
in an attempt to base this on the 'role' of a minion.

However, this implementation is flawed: instead of matching on the
minion ID in the `metalk8s.nodes` structure, it matches on the existence
of *any* minion in a specific role in the whole cluster, so de-facto all
minions get the same computed Pillar data.

The fix is trivial: instead of matching on `*` in the Pillar glob match,
only match on `grains.id` in `metalk8s.nodes`.

See: #1010
See: https://github.com/scality/metalk8s/pull/1010
See: 13485d2141d2ab46bbc375dd0d9e9e3a1a3be965