Define basic 'file.managed' wrapper state
We want to compute config file digests at runtime, so that the Salt
require/watch ordering rules can kick in first.
A quick solution was to wrap 'file.managed' inside a custom
'metalk8s.static_pod_managed' state, which computes a list of SHA256
digests from a list of file paths, and forwards it as a string to the
Pod manifest Jinja template.

Issue: GH-937