build/image: use skopeo instead of docker
In this PR we replace (most of) the use of docker by skopeo.
This concerns the creation of both local and remote images.

One remaining use of docker is to generate the `pause` image, because
`ctr cri load` can only load images saved by docker.
We will tackle this use case in a dedicated commit so that we can use
skopeo here as well.

Note that the images are no longer saved as tar archive, but as flat
directory (which will allow us to deduplicate the layers in a later
commit), thus the `Image` class can no longer inherits from the
`FileTarget` classe (because it doesn't produce a single file as
output).

This is not an issue as all image but one (`pause` again) don't rely on
this inheritance. For the last one, we introduce the `RemoteTarImage`
class.

Refs: #1049
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>