Document and use pytest markers
We document the existing @ci, @local, and @post markers for feature
files. One can run "pytest --markers" to see a list of registered
markers.
We use the "-m local" filter by default in tox environments, and switch
to "-m ci" when running tests in the CI.
We also introduce a @slow marker, and apply it to the "Bootstrap"
feature.
We then split CI test execution in two, running with "-m 'not slow'"
first, then "-m slow". It should help providing feedback faster to
developers.