build/salt: generate nginx config for registry
Since the `static-container-registry` tool is written in Python, we can
import it and call the right function instead of calling the tool as a
shell command, which avoid:
- handling stdout redirection (the tool can't write into a file)
- a fork to spawn a subcommand

Note that we have to use `importlib`, because we can't use the `import`
statement with names containing hyphens.

The call to `static-container-registry` is wrapped into a class that
inherits `targets.FileTarget` so that we can use it in the
`targets.FileTree`.

Refs: #1049
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>