build: get rid of `FileTarget` abstraction
In the long run, this has shown itself to be a wrong one that forced me
to jump through hoops (`RemoveTarImage`) so let's get rid of it and use
another approach.

The role of this class was to enforce targets with a single output file
(in order to use them in the `FileTree` class): let's add
`_get_destination` free function in `file_tree.py` that will dynamically
check if the target respect our contract (this will fail at the
"compile" time of the build, not during the build so it's fine).

As a side effect, we can now get rid of the ugly `RemoveTarImage`.

Refs: #1049
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>