charts: quote image name in MetalLB templates
During `helm template`, we splice a Jinja macro into the container image
repository value. When not quoted properly, later parsing of the
resulting YAML to fixup namespace information and such fails, because
the result is (prior to Jinja rendering) not valid YAML.

Simply quoting the YAML string value properly solves this.