build/images: delete intermediate containers
When building images, we sometimes spawn intermediate containers and
because those are not cleaned up, they accumulate and waste more and
more spaces.

We solve this problem by enabling the removal of the intermediate
containers (which is now the default behavior or the `docker build`
command anyway), in fact we even force the removal even in case of
failure (by using `forcerm` instead of `rm`).

Closes: #1164
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>