ci: deploy the SSH key pair for expansion
We deploy the public key on node1 and the private key on the bootstrap
node, so that the key pair can be used for cluster expansion.
Note that since we can't connect as root, we have to use two steps to
copy the private key in the right location (/etc/metalk8s/pki/).

This is a temporary workaround until we do it TheRightWay™ by using
Terraform.

Refs: #982, #1222
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>