tests: don't resolve hostname in single-node tests
In the single-node test environment, we only have the bootstrap node and
we force its hostname to "bootstrap" so there is no need to resolve it.
Moreover, trying to do so ends up with the tests failures (my guess it
that when giving an `ssh_config` set to `None` testinfra/paramiko still
try to use SSH, but fallback on a default configuration that won't work.

Given the fact that `ssh_config` is not set in the single-node test
environment, we use its value to know if we need to resolve the hostname
or not.

Refs: #982
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>