build/image: fix cleanup condition
When we save on disk, we use `skopeo` and it creates a file tree that
should be cleaned up with a recursive `rm`.
Otherwise, we simply touch a file and using `True` is enough to let
`doit` do the cleanup.

We simply have to  swap the branches to fix the issue.

Closes: #1373
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>