vagrant: fix OS comparison
During the review of the previous PR, we replaced the hash keys that
were strings by symbols, which is more idiomatic in Ruby (it's also more
efficient) We updated the code accordingly (hash lookup) but forgot
to fix the comparison.

As a consequence, the condition was always false and Python was
never installed on Ubuntu.

Using a symbol literal instead of a string literal is enough to fix the
issue.