salt: Remove specific node taint before uncordon
The commit ensures that we remove `node.kubernetes.io/unschedulable` node taints before
uncordon.
Initially, we did not remove this particular node taints before marking
them schedulable.
Now, to support downgrade we need remove this taint during the
uncordon process.

Source of new taint: With version 1.12 of k-c-m when cordoning a node
k-c-m adds this taints to the node