Store all image and package versions in buildchain
We define a single source of version information for the whole project,
and use Python to define such a file to easily read it from other
locations, such as tests or docs.

This module will be used throughout the buildchain, but should never
contain anything build-specific (such information can be located in the
relevant parts of other buildchain modules).

Issue: #640