storage-operator/salt: add a salt package
This new package provides a client to interact with the Salt API.
Since our reconciler will have to interact with Salt, we add a Salt
client as a new field.

The Salt client get the Salt master address from the environment
variable `METALK8S_SALT_MASTER_ADDRESS` and fallback to a sane default
(i.e. a default that will work if the operator runs in the same
namespace as the salt-master pod, which should be the case) if it's not
provided.
Using an environment variable allows us to use the operator even when
it's deployed outside of the cluster (which is the case in development
mode).

Example
=======

1. Get the IP of the salt-master (from the bootstrap node):
   => kubectl -n kube-system describe svc salt-master | grep 'Endpoints'
2. Run the operator locally (from your computer):
   => export METALK8S_SALT_MASTER_ADDRESS=http://<SALT-MASTER-IP>
   => operator-sdk up local

Refs: #1412
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>