storage-operator/salt: add an helper for POST
Note that we always set `Content-Type` and `Accept` HTTP headers because
we're sending (and expect receiving[1]) JSON payloads.

The `headers` parameters can be used to specify extra HTTP headers (like
`X-Auth-Token` to specify the Salt authentication token).

[1]: but in some case, like getting a 401 on authentication, Salt API
will happily returns an HTML payload even though you requested JSON…
(that's why is the status code is not 200, we don't try to decode the
payload).

Refs: #1412
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>