storage-operator/salt: handle token invalidation
In some case, a valid token (i.e. not expired) may be rejected by the
server[1] and we should gracefully handle this case by trying to
re-authenticate to get a new token and then retry the request.

As a side effect, now that our client has a good handling of the token
management, the user of our client no longer need to authenticate
themselve (thus, the method is now private).

[1]: one case where it happen is when the Salt API server is restarted,
in that case all the tokens are lost (for security reason they are
stored in RAM, not on disk) and thus are no longer accepted.

Refs: #1412
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>