Pull images by version and check digest
Previously, we pulled by digest and applied the declared version as a
tag. This implied that updating the declared version only changed the
tag of the image, without actually changing the image itself.

We now make sure to pull the desired version with both Docker and
Skopeo, checking the SHA256 digest of the result against what we specify
in `buildchain.versions`.

Issue: #640