storage-operator/salt: add support for ServiceAccount token
Until now, we were using the default hardcoded credentials but ideally
we should use the ServiceAccount token to authenticate.

Since we want to Salt API client to be independant to Kubernetes, the
creds lookup is done in the Volume controller.

Note that we keep a fallback on the default creds for now (because in
local dev mode, you don't have a ServiceAccount).

Refs: #1412
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>