salt: use `kube-system` annotation during upgrade/downgrade
This commit makes it possible for us to use the kube-system annotation.
Initially, we passed the destination version directly to the bash upgrade/downgrade
script.
With this commit, we now store and update the value of the annotation
each time we initialize an upgrade or downgrade. The new annotation value comes
from the input argument(--destination-version) to our upgrade/downgrade script.