Generate versions list in Salt tree
In order to read the list of versions defined in the buildchain from
Salt formulas, we need to render them into a static file (we chose
JSON, since YAML requires adding a dependency, and both are the easiest
to read from).

This is then read by `map.jinja` as an extension to the existing
`defaults.yaml`.

Issue: #640