Fix task callables signatures and type hints
In an attempt to clean the signature of the `docker_command` task
callables, we used the `Optional[TaskError]` return type hint.
However, this expressely required us to add a `return None` statement
for `mypy`, and disable a `pylint` rule to do so.

Instead, we use the return type hint `None` and don't `return` from such
callables, leaving the `Optional[TaskError]` to the `task_error`
decorator.

Fixes: #1489
See: #1477