storage-operator/salt: add PrepareVolume
Add a method to call the Salt state that will prepare the volume. Since
this call is expected to take a while to execute, we spawn it
asynchronously (otherwise we take the risk to block the reconciliation
loop, which we clearly want to avoid).

Note that the name is subject to change (because the Salt state is not
developped yet)

Refs: #1462
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>