storage-operator/salt: add PollJob
We add a new public method to our Salt client: PollJob.
This method will be used to poll the state of a job that was spawned
asynchronously.

Note that since we want to make the difference between a failure that is
transient (the polling failed) vs a failure that is definitive (the job
executed and ended up with an error), we introduce a new type of error
`AsyncJobFailed` to represent the latter.

That way, the reconciliation loop can decide to reschedule or move to a
failed state accordingly.

Refs: #1462
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>