storage-operator/salt: add timeout on Salt request
For synchronous execution, using `local`, we pass the timeout in the
request payload.

For job polling, we add a timeout parameter to the URL.

Note that to add timeout in the payload, we need to change the type of
the hasmap used (we no longer use only string as value).

Refs: #1462
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>