storage-operator/volume: cleanup deleted object handling
When a volume without finalizers (i.e. not yet/longer associated to a
PV) is deleted, the reconciliation loop may be called (triggered by the
deletion event) but the volume is already deleted (because no finalizer
were retaining it).

The case was properly handled, but the comment wasn't really explicit.

Also remove `instance` to `volume` for the sake of explicitness.

Refs: #1462
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>